src/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
supabase
pyjwt[crypto]
httpx
python-multipart
pymupdf
openai
celery
redis